server/requirements.txt
# fastapi==0.104.1
# uvicorn[standard]==0.24.0
# pandas==2.1.3
# scikit-learn==1.3.2
# python-multipart==0.0.6
# python-dotenv==1.0.0
# numpy==1.26.2
# pydantic==2.5.0


fastapi
uvicorn
pandas
scikit-learn
numpy
python-multipart
python-dotenv
pydantic